(PUP-5222) Add upgrading the puppet-agent package in Puppet 4+
Adds support for upgrading the puppet-agent package after install.
Upgrades of Puppet 4+ agents will only occur if a specific version is
specified.

Also adds a collection parameter to specify which Puppet Collection to
use.
