+++ manifests/install.pp
@@ -2,10 +2,21 @@
 #
 # This class is called from puppet_agent for install.
 #
-class puppet_agent::install {
+class puppet_agent::install (
+  $version
+) {
   assert_private()
 
-  package { $::puppet_agent::package_name:
-    ensure => present,
+  if $::puppet_agent::params::_windows_client {
+    class { '::puppet_agent::install::windows':
+      version => $version,
+    }
+    contain '::puppet_agent::install::windows'
+  }
+  else {
+    class { '::puppet_agent::install::default':
+      version => $version,
+    }
+    contain '::puppet_agent::install::default'
   }
 }
