+++ manifests/osfamily/debian.pp
@@ -28,7 +28,7 @@ class puppet_agent::osfamily::debian {
       "Acquire::http:::proxy::${source_host} DIRECT;",
     ]
 
-    apt::setting { 'conf-pc1_repo':
+    apt::setting { 'conf-pc_repo':
       content  => $_apt_settings.join(''),
       priority => 90,
     }
@@ -57,20 +57,20 @@ class puppet_agent::osfamily::debian {
   }
 
 
-  apt::source { 'pc1_repo':
+  apt::source { 'pc_repo':
     location => $source,
-    repos    => 'PC1',
+    repos    => $::puppet_agent::collection,
     key      => {
       'id'     => '47B320EB4C7C375AA9DAE1A01054B7A24BD6EC30',
       'server' => 'pgp.mit.edu',
     },
-    notify   => Notify['pc1_repo_force'],
+    notify   => Notify['pc_repo_force'],
   }
 
   # apt_update doesn't inherit the future class dependency, so it
   # can wait until the end of the run to exec. Force it to happen now.
-  notify { 'pc1_repo_force':
-      message => 'forcing apt update for pc1_repo',
+  notify { 'pc_repo_force':
+      message => 'forcing apt update for pc_repo',
       require => Exec['apt_update'],
   }
 }
