+++ spec/classes/puppet_agent_spec.rb
@@ -1,6 +1,7 @@
 require 'spec_helper'
 
 describe 'puppet_agent' do
+
   context 'supported operating systems' do
     on_supported_os.each do |os, facts|
       context "on #{os}" do
@@ -10,17 +11,19 @@ describe 'puppet_agent' do
           it { expect { is_expected.to contain_package('puppet_agent') }.to raise_error(Puppet::Error, /upgrading requires Puppet 3.8/) }
         else
           [{}, {:service_names => []}].each do |params|
-            context "puppet_agent class without any parameters" do
+            context "puppet_agent class with parameters #{params}" do
               let(:params) { params }
 
               it { is_expected.to compile.with_all_deps }
 
               it { is_expected.to contain_class('puppet_agent') }
               it { is_expected.to contain_class('puppet_agent::params') }
-              if Puppet.version < "4.0.0"
-                it { is_expected.to contain_class('puppet_agent::prepare') }
-                it { is_expected.to contain_class('puppet_agent::install').that_comes_before('puppet_agent::service') }
-                it { is_expected.to contain_class('puppet_agent::service') }
+              it { is_expected.to contain_class('puppet_agent::prepare') }
+              it { is_expected.to contain_class('puppet_agent::install') }
+              it { is_expected.to contain_package('puppet-agent').with_ensure('present') }
+
+              if os != 'windows'
+                it { is_expected.to contain_class('puppet_agent::service').that_requires('puppet_agent::install') }
 
                 if params[:service_names].nil?
                   it { is_expected.to contain_service('puppet') }
@@ -29,11 +32,9 @@ describe 'puppet_agent' do
                   it { is_expected.to_not contain_service('puppet') }
                   it { is_expected.to_not contain_service('mcollective') }
                 end
-                it { is_expected.to contain_package('puppet-agent').with_ensure('present') }
               else
                 it { is_expected.to_not contain_service('puppet') }
                 it { is_expected.to_not contain_service('mcollective') }
-                it { is_expected.to_not contain_package('puppet-agent') }
               end
             end
           end
